refactor(core): split value journal tests by behavior
