+++ src/core/src/state_journal/tests/mod.rs
@@ -1,0 +1,19 @@
+#![cfg(all(
+    feature = "f64",
+    feature = "record",
+    feature = "tuple",
+    feature = "map",
+    feature = "set",
+    feature = "table",
+    feature = "enum",
+    feature = "matrixd",
+    feature = "vectord"
+))]
+
+mod borrow_conflicts;
+mod collections;
+mod collisions;
+mod delta;
+mod nested;
+mod scalar;
+mod support;
